Turn on warnings (GCC_WARN_INHIBIT_ALL_WARNINGS was previously YES, so we weren't seeing any before! :-). The code should run the same. I left some warnings in the build until I can figure out how to best handle them. I don't want to turn these warnings off yet so I'm reminded to come up with a better solution...

+++ OhmPlayer3/OhmPlayer3/MutablePlaylist.m
@@ -108,9 +108,9 @@
 {
     if (destinationIndex != sourceIndex)
     {        
-        id obj = [mutableArray objectAtIndex:sourceIndex];
+        id obj = [mutableArray objectAtIndex:(NSUInteger)sourceIndex];
         
-        [mutableArray removeObjectAtIndex:sourceIndex];
+        [mutableArray removeObjectAtIndex:(NSUInteger)sourceIndex];
         
         if (destinationIndex >= (NSInteger)[mutableArray count])
         {
@@ -118,7 +118,7 @@
         }
         else
         {
-            [mutableArray insertObject:obj atIndex:destinationIndex];
+            [mutableArray insertObject:obj atIndex:(NSUInteger)destinationIndex];
         }
     }
 }
